pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "flagguard"
version = "0.1.0"
description = "AI Feature Flag Conflict Analyzer"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "Laxmi Ranjan", email = "laxmiranjan444@gmail.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
keywords = ["feature-flags", "static-analysis", "launchdarkly"]

dependencies = [
    "pyyaml>=6.0",
    "z3-solver>=4.12",
    "tree-sitter>=0.21",
    "tree-sitter-python>=0.21",
    "tree-sitter-javascript>=0.21",
    "networkx>=3.0",
    "click>=8.0",
    "rich>=13.0",
    "gradio>=4.0",
    "ollama>=0.1",
    "chromadb>=1.5.0",
    "sentence-transformers>=5.2.2",
    "matplotlib>=3.7.0",
    "plotly>=5.0.0",
    "sqlalchemy>=2.0.46",
    "alembic>=1.18.4",
    "psycopg2-binary>=2.9.11",
    "python-jose[cryptography]>=3.5.0",
    "passlib[bcrypt]>=1.7.4",
    "email-validator>=2.3.0",
    "uvicorn>=0.40.0",
    "requests>=2.32.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.1",
    "mypy>=1.0",
]

[project.scripts]
flagguard = "flagguard.cli.main:cli"

[project.urls]
Homepage = "https://github.com/laxmi2577/flagguard"
Repository = "https://github.com/laxmi2577/flagguard"

[tool.hatch.build.targets.wheel]
packages = ["src/flagguard"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
]
